Fix HTTP Auth failure with old (pre UTF-8) servers
Some servers, like really old SabreDAV based servers, seem to choke on
password if we pass a bytes sequence instead of a string to the
"requests" libraries HTTPDigestAuth object.

This Fix retries authentication in case of authentication failure with
the same password as plain string (which, in Python3 is anyway UTF-8).